fix admin acces only

+++ ajax/deleteJapan.php
@@ -1,22 +1,22 @@
 <?php
 
-require_once '../model/frontend.php';
-require_once '../model/backend.php';
+if ($_SESSION['connect'] === 'OK' && $_SESSION['admin'] === 1) {
+    require_once '../model/frontend.php';
+    require_once '../model/backend.php';
 
-deleteAllKanjiForJaponais($_POST['id']);
-deleteAllGroupeForJaponais($_POST['id']);
-$anglais = listAnglaisToJaponais($_POST['id']);
-foreach ($anglais as $word) {
-    deleteAnglais($word['id']);
-    supprAnglais($word['id']);
-}
-$francais = listFrancaisToJaponais($_POST['id']);
-foreach ($francais as $word) {
-    deleteFrancais($word['id']);
-    supprWord($word['id']);
-}
-$deleteJaponais = supprJaponais($_POST['id']);
+    deleteAllKanjiForJaponais($_POST['id']);
+    deleteAllGroupeForJaponais($_POST['id']);
+    $anglais = listAnglaisToJaponais($_POST['id']);
+    foreach ($anglais as $word) {
+        deleteAnglais($word['id']);
+        supprAnglais($word['id']);
+    }
+    $francais = listFrancaisToJaponais($_POST['id']);
+    foreach ($francais as $word) {
+        deleteFrancais($word['id']);
+        supprWord($word['id']);
+    }
+    $deleteJaponais = supprJaponais($_POST['id']);
 
-if ($deleteJaponais == true) {
-    echo 'SUCCESS';
-}
+    if ($deleteJaponais == true) echo 'SUCCESS';
+} else echo 'WARNING';
